Actualizando Bloque  01
Listo para funcionalidad con frontend

+++ Sln_VetManadaWoof/src/main/java/com/vet/manadawoof/controller/EmpresaRestController.java
@@ -2,13 +2,14 @@ package com.vet.manadawoof.controller;
 
 import com.vet.manadawoof.entity.EmpresaEntity;
 import com.vet.manadawoof.service.EmpresaService;
+import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
-@CrossOrigin(origins ="http://localhost:5173")
+@CrossOrigin(origins = "http://localhost:5173")
 @RestController
 @RequestMapping("/api/empresas")
 @RequiredArgsConstructor
@@ -22,14 +23,16 @@ public class EmpresaRestController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<EmpresaEntity> obtenerEmpresa(@PathVariable Long id) {
+    public ResponseEntity<EmpresaEntity> obtenerEmpresa(@PathVariable Integer id)
+    {
         EmpresaEntity empresa = service.obtenerEmpresa(id);
         return empresa != null ? ResponseEntity.ok(empresa) : ResponseEntity.notFound().build();
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<String> actualizarEmpresa(@PathVariable Long id,
-                                                    @RequestBody EmpresaEntity empresa) {
+    public ResponseEntity<String> actualizarEmpresa(
+            @PathVariable Integer id,
+            @RequestBody EmpresaEntity empresa) {
         empresa.setId(id);
         return ResponseEntity.ok(service.actualizarEmpresa(empresa));
     }
